refactor(build): Refactor tests for unit/int/bench
